feat: add file sharing, add view of available to me files, add removing shared files

+++ src/main/java/com/cfs/cloudfilestorage/service/storage/Impl/command/DBRemoveCommand.java
@@ -2,6 +2,7 @@ package com.cfs.cloudfilestorage.service.storage.Impl.command;
 
 import com.cfs.cloudfilestorage.dto.StorageEntity;
 import com.cfs.cloudfilestorage.repository.ItemRepository;
+import com.cfs.cloudfilestorage.repository.PersonRepository;
 import com.cfs.cloudfilestorage.service.person.AuthorizedPersonService;
 import com.cfs.cloudfilestorage.service.storage.StorageCommand;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -10,10 +11,13 @@ public class DBRemoveCommand extends StorageCommand {
 
     private final ItemRepository itemRepository;
     private final AuthorizedPersonService authorizedPersonService;
+    private final PersonRepository personRepository;
 
     public DBRemoveCommand(ItemRepository itemRepository,
+                           PersonRepository personRepository,
                            AuthorizedPersonService authorizedPersonService){
         this.itemRepository = itemRepository;
+        this.personRepository = personRepository;
         this.authorizedPersonService = authorizedPersonService;
     }
 
@@ -37,6 +41,8 @@ public class DBRemoveCommand extends StorageCommand {
 
         person.getAvailableItems().remove(item);
 
-        itemRepository.delete(item);
+        if(item.getOwnerEmail().equals(person.getEmail()))
+            itemRepository.delete(item);
+        personRepository.save(person);
     }
 }
